requirements.txt
Django==4.2.11
djangorestframework==3.14.0
psycopg[binary,pool]==3.1.12; implementation_name != 'pypy'
redis[hiredis]==5.0.1
